Split implementation into several files.
